backend/requirements.txt
sqlalchemy==2.0.23
pymysql==1.1.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
openai==1.3.0
